Use Renderers for update and drawing operations
