--- core/src/main/java/com/alamkanak/weekview/Constants.kt
@@ -1,7 +1,0 @@
-package com.alamkanak.weekview
-
-internal object Constants {
-    const val DAY_IN_MILLIS = 1000L * 60L * 60L * 24L
-    const val MINUTES_PER_HOUR = 60f
-    const val UNINITIALIZED = -1f
-}
