+++ core/src/main/java/com/alamkanak/weekview/Renderers.kt
@@ -11,9 +11,15 @@ internal interface Drawer {
     fun draw(canvas: Canvas)
 }
 
-// TODO
+internal interface DateFormatterDependent {
+    fun onDateFormatterChanged(formatter: DateFormatter)
+}
+
+internal interface TimeFormatterDependent {
+    fun onTimeFormatterChanged(formatter: TimeFormatter)
+}
 
 internal interface Renderer {
-    fun onSizeChanged(width: Int, height: Int)
+    fun onSizeChanged(width: Int, height: Int) = Unit
     fun render(canvas: Canvas)
 }
