+++ sample/src/main/res/layout/activity_basic.xml
@@ -8,43 +8,65 @@
 
     <include layout="@layout/view_toolbar" />
 
-    <com.alamkanak.weekview.WeekView
-        android:id="@+id/weekView"
+    <FrameLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:adaptiveEventTextSize="true"
-        app:allDayEventTextSize="13sp"
-        app:columnGap="1dp"
-        app:eventCornerRadius="4dp"
-        app:eventPaddingHorizontal="6dp"
-        app:eventPaddingVertical="2dp"
-        app:eventTextColor="@android:color/white"
-        app:headerRowBottomShadowColor="@color/light_gray"
-        app:headerRowBottomShadowRadius="4dp"
-        app:headerRowPadding="12dp"
-        app:hourHeight="60dp"
-        app:nowLineColor="@color/accent"
-        app:nowLineDotColor="@color/accent"
-        app:nowLineDotRadius="5dp"
-        app:nowLineStrokeWidth="1dp"
-        app:numberOfVisibleDays="3"
-        app:overlappingEventGap="1dp"
-        app:showCurrentTimeFirst="true"
-        app:showHeaderRowBottomShadow="true"
-        app:showNowLine="true"
-        app:showNowLineDot="true"
-        app:showTimeColumnSeparator="true"
-        app:showWeekNumber="true"
-        app:singleDayHorizontalMargin="8dp"
-        app:timeColumnBackgroundColor="#ffffffff"
-        app:timeColumnPadding="8dp"
-        app:timeColumnSeparatorColor="@color/header_bottom_line"
-        app:timeColumnSeparatorStrokeWidth="1dp"
-        app:timeColumnTextColor="@color/toolbar_text"
-        app:timeColumnTextSize="12sp"
-        app:todayHeaderTextColor="@color/accent"
-        app:weekNumberBackgroundColor="@color/dark_gray"
-        app:weekNumberBackgroundCornerRadius="@dimen/corner_radius"
-        app:weekNumberTextColor="@android:color/white" />
+        android:layout_height="0dp"
+        android:layout_weight="1">
+
+        <com.alamkanak.weekview.WeekView
+            android:id="@+id/weekView"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            app:adaptiveEventTextSize="true"
+            app:allDayEventTextSize="13sp"
+            app:columnGap="1dp"
+            app:eventCornerRadius="4dp"
+            app:eventPaddingHorizontal="6dp"
+            app:eventPaddingVertical="2dp"
+            app:eventTextColor="@android:color/white"
+            app:headerRowBottomShadowColor="@color/light_gray"
+            app:headerRowBottomShadowRadius="4dp"
+            app:headerRowPadding="12dp"
+            app:hourHeight="60dp"
+            app:nowLineColor="@color/accent"
+            app:nowLineDotColor="@color/accent"
+            app:nowLineDotRadius="5dp"
+            app:nowLineStrokeWidth="1dp"
+            app:numberOfVisibleDays="3"
+            app:overlappingEventGap="1dp"
+            app:showCurrentTimeFirst="true"
+            app:showHeaderRowBottomShadow="true"
+            app:showNowLine="true"
+            app:showNowLineDot="true"
+            app:showTimeColumnSeparator="true"
+            app:showWeekNumber="true"
+            app:singleDayHorizontalMargin="8dp"
+            app:timeColumnBackgroundColor="@android:color/white"
+            app:timeColumnPadding="8dp"
+            app:timeColumnSeparatorColor="@color/header_bottom_line"
+            app:timeColumnSeparatorStrokeWidth="1dp"
+            app:timeColumnTextColor="@color/toolbar_text"
+            app:timeColumnTextSize="12sp"
+            app:todayHeaderTextColor="@color/accent"
+            app:weekNumberBackgroundColor="@color/dark_gray"
+            app:weekNumberBackgroundCornerRadius="@dimen/corner_radius"
+            app:weekNumberTextColor="@android:color/white" />
+
+        <FrameLayout
+            android:id="@+id/blockingProgressIndicator"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:background="#80FFFFFF"
+            android:visibility="gone"
+            tools:visibility="visible">
+
+            <ProgressBar
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center" />
+
+        </FrameLayout>
+
+    </FrameLayout>
 
 </LinearLayout>
